Move api-job code into job routes file.
